Move verification data to the public package
