+++ src/main/java/org/mockito/internal/verification/VerificationDataImpl.java
@@ -6,7 +6,7 @@ package org.mockito.internal.verification;
 
 import org.mockito.internal.invocation.InvocationMatcher;
 import org.mockito.internal.stubbing.InvocationContainer;
-import org.mockito.internal.verification.api.VerificationData;
+import org.mockito.verification.VerificationData;
 import org.mockito.invocation.Invocation;
 
 import static org.mockito.internal.exceptions.Reporter.cannotVerifyToString;
